+++ pkg/cli/internal/cmd/setup/setup.go
@@ -41,7 +41,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
 	operatorv1alpha1 "github.com/kubermatic/kubecarrier/pkg/apis/operator/v1alpha1"
-	"github.com/kubermatic/kubecarrier/pkg/cli/internal/cmd/preflight"
+	"github.com/kubermatic/kubecarrier/pkg/cli/internal/cmd/preflight/checkers"
 	"github.com/kubermatic/kubecarrier/pkg/cli/internal/spinner"
 	"github.com/kubermatic/kubecarrier/pkg/internal/reconcile"
 	"github.com/kubermatic/kubecarrier/pkg/internal/resources/operator"
@@ -95,7 +95,7 @@ func runE(conf *rest.Config, log logr.Logger, cmd *cobra.Command) error {
 
 	s := wow.New(cmd.OutOrStdout(), spin.Get(spin.Dots), "")
 	startTime := time.Now()
-	if err := preflight.RunChecks(conf, s, startTime, log); err != nil {
+	if err := checkers.RunChecks(conf, s, startTime, log); err != nil {
 		return err
 	}
 	// Get a client from the configuration of the kubernetes cluster.
